Add files via upload
This is mini project about the football club management and scouting networks. This project deals with the trading of the players between different clubs , purchasing and selling them for club wedges , recruiting new scouts by the scouts of the respective football clubs. The whole transactions are visible to the club owner (Admin) who is responsible for the club wedges. This project has included PHP,HTML,CSS for development and is basically a Database Management Project

1.Download all the files in the repository
2.Download Xampp server to your system
3.Turn apache & MYSQL server on and search "localhost/dashboard" in your local browser
4.Create new Database there and import the file "proj (3).sql" from the repository.
5.Now copy all the files except proj (3).sql and paste them into a folder naming say 'abc'
6.Now go to your browser & search "localhost/abc" the project will run automatically.
7. There you will find 2 login buttons,select either one of them and fill the credentials according to the entries in the proj (3).sql .